fix(files): setup cannot be in mixins
